+++ opentelemetry-otlp/src/exporter/http/mod.rs
@@ -1047,7 +1047,7 @@ mod tests {
             SpanData {
                 span_context,
                 parent_span_id: SpanId::from(0),
-                parent_span_context: None,
+                parent_span_is_remote: false,
                 span_kind: SpanKind::Internal,
                 name: Cow::Borrowed("test_span"),
                 start_time: SystemTime::UNIX_EPOCH,