+++ opentelemetry-sdk/src/trace/span_processor.rs
@@ -968,7 +968,7 @@ mod tests {
         let unsampled = SpanData {
             span_context: SpanContext::empty_context(),
             parent_span_id: SpanId::INVALID,
-            parent_span_context: None,
+            parent_span_is_remote: false,
             span_kind: SpanKind::Internal,
             name: "opentelemetry".into(),
             start_time: opentelemetry::time::now(),
@@ -1130,7 +1130,7 @@ mod tests {
         SpanData {
             span_context: SpanContext::empty_context(),
             parent_span_id: SpanId::INVALID,
-            parent_span_context: None,
+            parent_span_is_remote: false,
             span_kind: SpanKind::Internal,
             name: name.to_string().into(),
             start_time: opentelemetry::time::now(),