+++ opentelemetry-zipkin/src/exporter/model/span.rs
@@ -154,7 +154,7 @@ mod tests {
                     Default::default(),
                 ),
                 parent_span_id: SpanId::from(1),
-                parent_span_context: None,
+                parent_span_is_remote: false,
                 span_kind: SpanKind::Client,
                 name: "".into(),
                 start_time: now(),